refactor: removes deprecated methods
